+++ app/src/main/java/eu/faircode/email/FragmentOptions.java
@@ -133,6 +133,7 @@ public class FragmentOptions extends FragmentBase {
             "send_pending",
             "startup", "cards", "beige", "tabular_card_bg", "shadow_unread", "shadow_highlight",
             "portrait2", "portrait2c", "portrait_min_size", "landscape", "landscape_min_size",
+            "column_width",
             "nav_count", "nav_unseen_drafts", "navbar_colorize",
             "indentation", "group_category", "date", "date_fixed", "date_bold", "threading", "threading_unread",
             "highlight_unread", "highlight_color", "color_stripe", "color_stripe_wide",